+++ install.ts
@@ -1,22 +1,32 @@
-import { join, parse } from "https://deno.land/std@0.66.0/path/mod.ts";
+import { join, parse, resolve } from "https://deno.land/std@0.66.0/path/mod.ts";
 
 const url = import.meta.url;
 
 function buildURL(target: string): string {
   const { dir } = parse(url);
-  return new URL(join(dir, target)).href;
+  if (url.startsWith("https://")) {
+    return new URL(join(dir, target)).href;
+  } else {
+    return resolve(dir.replace("file://", ""), target);
+  }
 }
 
 // Cache all deps
 
 // Main deps
 await Deno.run({
-  cmd: ["deno", "cache", "--unstable", `${buildURL("./deps.ts")}`],
+  cmd: ["deno", "cache", "--unstable", "--reload", `${buildURL("./deps.ts")}`],
 }).status();
 
 // Terser
 await Deno.run({
-  cmd: ["deno", "cache", "--unstable", `${buildURL("./plugin/terserTransform/deps.ts")}`],
+  cmd: [
+    "deno",
+    "cache",
+    "--unstable",
+    "--reload",
+    `${buildURL("./plugin/terserTransform/deps.ts")}`,
+  ],
 }).status();
 
 // Unstable plugin deps
@@ -26,6 +36,7 @@ await Deno.run({
     "cache",
     "--unstable",
     "--no-check", // remove when https://github.com/denoland/deno/issues/7145 is resolved
+    "--reload",
     `${buildURL("./plugin/deps.ts")}`,
   ],
 }).status();