+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0](https://github.com/denofn/denopack/compare/0.6.0..0.7.0) - 2020-08-22
+
+### Added
+
+- Change to custom install script
+
+### Changed
+
+- Switched out custom filter function in cssBundle for official Rollup createFilter function
+
 ## [0.6.0](https://github.com/denofn/denopack/compare/0.5.0..0.6.0) - 2020-08-18
 
 ### Added